% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{as.character.jsonString}
\alias{as.character.jsonString}
\title{JSON string to character string}
\usage{
\method{as.character}{jsonString}(jsonstring)
}
\arguments{
\item{jsonstring}{JSON string}
}
\value{
A character string.
}
\description{
Convert a JSON string to a character string.
}
