% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonAddProperty}
\alias{jsonAddProperty}
\title{Add new property}
\usage{
jsonAddProperty(jsonstring, key, value)
}
\arguments{
\item{jsonstring}{a JSON string representing an object}

\item{key}{a character string, the key of the new property}

\item{value}{a JSON string, the value of the new property}
}
\value{
A JSON string.
}
\description{
Add a new property to a JSON string.
}
\examples{
jsonAddProperty("{\"a\":[1,2,3],\"b\":\"hello\"}", "c", "[1,2]")
}
