% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonAt}
\alias{jsonAt}
\title{Access an element in a JSON string}
\usage{
jsonAt(jsonstring, path)
}
\arguments{
\item{path}{path to the element to extract; either a character vector of 
keys, an integer vector of indices, or a list made of keys and indices}

\item{jsonString}{a JSON string representing an array or an object}
}
\value{
A JSON string.
}
\description{
Extract an element in a JSON string by giving a path of keys or 
  indices.
}
\examples{
jstring <- "[1,[\"a\",99],{\"x\":[2,3,4],\"y\":42}]"
jsonAt(jstring, 1)
jsonAt(jstring, list(2, "x"))
jsonAt(jstring, list(2, "z"))
}
