% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonErase}
\alias{jsonErase}
\title{Erase property or element}
\usage{
jsonErase(jsonstring, at)
}
\arguments{
\item{jsonstring}{JSON string representing an object or an array}

\item{at}{either a character string, the key of the property to be erased, 
or an integer, the index of the array element to be erased}
}
\value{
A JSON string.
}
\description{
Erase an object property or an array element from a JSON string.
}
\examples{
jsonErase("{\"a\":[1,2,3],\"b\":\"hello\"}", "a")
}
