% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonHasKey}
\alias{jsonHasKey}
\title{Does key exist?}
\usage{
jsonHasKey(jsonString, key)
}
\arguments{
\item{jsonString}{a JSON string}

\item{key}{character string}
}
\value{
\code{TRUE} if the given key is present in the JSON string, 
  \code{FALSE} otherwise.
}
\description{
Checks whether a key is present in a JSON string.
}
\examples{
jsonHasKey("{\"a\":[1,2,3],\"b\":\"hello\"}", "b")
jsonHasKey("[1,2,3]", "a")
}
