% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonIs}
\alias{jsonIs}
\title{Check type of JSON string}
\usage{
jsonIs(jsonstring, type)
}
\arguments{
\item{jsonstring}{a JSON string}

\item{type}{the type to be checked, one of \code{"array"}, \code{"object"}, 
\code{"string"}, \code{"number"}, \code{"integer"}, \code{"null"}, 
\code{"boolean"}}
}
\value{
A logical value.
}
\description{
Check the type of a JSON string.
}
