% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonMerge}
\alias{jsonMerge}
\title{Merge JSON strings}
\usage{
jsonMerge(jsonstring1, jsonstring2)
}
\arguments{
\item{jsonstring1}{a JSON string}

\item{jsonstring2}{a JSON string}
}
\value{
A JSON string.
}
\description{
Merge two JSON strings.
}
\examples{
jstring1 <- jsonString("{\"a\":[1,2,3],\"b\":\"hello\"}")
jstring2 <- jsonString("{\"a\":[4,5,6],\"c\":\"goodbye\"}")
jsonMerge(jstring1, jstring2)
}
