% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonPush}
\alias{jsonPush}
\title{Append an element}
\usage{
jsonPush(jsonstring1, jsonstring2)
}
\arguments{
\item{jsonstring1}{JSON string representing an array}

\item{jsonstring2}{JSON string}
}
\value{
A JSON string.
}
\description{
Append an element to a JSON string.
}
\examples{
jstring1 <- jsonString("[1,2,3]")
jstring2 <- jsonString("{\"a\":\"hello\",\"b\":\"goodbye\"}")
jsonPush(jstring1, jstring2)
}
