% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonSize}
\alias{jsonSize}
\title{Number of elements}
\usage{
jsonSize(jsonstring)
}
\arguments{
\item{jsonstring}{a JSON string}
}
\value{
An integer.
}
\description{
Number of elements in a JSON string.
}
\examples{
jsonSize("{\"a\":[1,2,3],\"b\":\"hello\"}")
}
