% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\docType{class}
\name{jsonStrin-class}
\alias{jsonStrin-class}
\alias{jsonStrin}
\title{Class name}
\description{
Length class description
}
\section{Fields}{

\describe{
\item{\code{text}}{Text field description}

\item{\code{fooBar}}{Description of the fooBar field}
}}

\section{Methods}{

\describe{
\item{\code{getParameters()}}{Second method lengthy description.
\subsection{Parameters}{\itemize{
\item{\code{fooBar} Another fooBar parameter description.}
\item{\code{fooBaz} Yet another parameter description.}
}}
\subsection{Return Value}{Another return value}}

\item{\code{initialize(string)}}{Creates a JSON string.
\subsection{Parameters}{\itemize{
  \item{\code{string} a string representing a JSON object}
}}
\subsection{Return value}{An object of class \code{jsonString}.}}

\item{\code{runs(a, b)}}{Lengthy method description.
\subsection{Parameters}{\itemize{
  \item{\code{foo} Some foo parameter \code{code} here.}
  \item{\code{bar} Some bar parameter description.}
}}
\subsection{Return Value}{A return value}}
}}

