% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonType}
\alias{jsonType}
\title{Type of JSON string}
\usage{
jsonType(jsonstring)
}
\arguments{
\item{jsonstring}{a JSON string}
}
\value{
A character string indicating the type of the JSON string.
}
\description{
The type of a JSON string.
}
