% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{as.character.jsonString}
\alias{as.character.jsonString}
\title{JSON string to character string}
\usage{
\method{as.character}{jsonString}(jsonstring, pretty = FALSE)
}
\arguments{
\item{jsonstring}{JSON string}

\item{pretty}{logical value, whether to pretty-format the string}
}
\value{
A character string.
}
\description{
Convert a JSON string to a character string.
}
