% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{jsonString}
\alias{jsonString}
\title{JSON string}
\usage{
jsonString(string)
}
\arguments{
\item{string}{a character string}
}
\value{
A JSON string (external pointer).
}
\description{
Create a JSON string.
}
\examples{
jsonString("[1,[\"a\",99],{\"x\":[2,3,4],\"y\":42}]")
}
