% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{JsonString}
\alias{JsonString}
\title{Class of JSON strings objects}
\description{
Type the name of the class to see its methods
}
\section{Fields}{

\describe{
\item{\code{new}}{Constructor}

\item{\code{mult}}{Multiply by another Double object \itemize{
\item Parameter: other - The other Double object
\item Returns: product of the values
}}
}}

