% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jsonStrings.R
\name{print.jsonString}
\alias{print.jsonString}
\title{Print JSON string}
\usage{
\method{print}{jsonString}(x, pretty = getOption("jsonStrings.prettyPrint", FALSE), ...)
}
\arguments{
\item{x}{JSON string}

\item{pretty}{logical value, whether to pretty-print}

\item{...}{ignored}
}
\description{
Print a JSON string.
}
